infrastructure/requirements.txt
django
djangorestframework
flask_restful
psycopg2-binary
python-dotenv